client/python: move git helpers to git.py and add error handling in sourceCodeLocation facet
